Fix failing test cases
This is something I've got to figure out how to manage. Naturally, Brian doesn't run the test suite but when he makes changes it breaks tests. When I go to develop, I have to spend time fixing all the tests so I can start at a green state. In this case, it took 45 minutes to get everything in line.